added more to the Teensy code. Need to use Rumman's Teensy CAN tutorial to solve CAN problems

+++ ESC Controller/ESC_Controller_CAN/ESC_Controller_CAN/sensors.h
@@ -7,25 +7,25 @@ class Sensors {
 public:
 
   Sensors() {};
-  uint16_t mcurrent(int curr_sens_pin);
-  uint16_t mspeed(int enc1_count, int enc_inc, int enc1_avg, int enc1_val, int enc1_hold, int count, int wheel_diam, int gear_rat, uint16_t runtime, uint16_t speedtimer);
-  uint16_t pinger();
+  uint8_t mcurrent(int curr_sens_pin);
+  uint8_t mspeed(int enc1_count, int enc_inc, int enc1_avg, int enc1_val, int enc1_hold, int count, int wheel_diam, int gear_rat, uint16_t runtime, uint16_t speedtimer);
+  uint8_t pinger();
 
 private:
   int enc1;
   int curr_sens_pin;
-  int16_t speedtimer;
+  int8_t speedtimer;
   int enc1_val;
   int enc_inc;
   int enc1_count;
   int enc1_avg;
   int count;
   int enc1_hold;
-  uint16_t runtime;
+  uint8_t runtime;
   int wheel_diam;
   int gear_rat;
-  uint16_t curr_val;
-  uint16_t velocity;
+  uint8_t curr_val;
+  uint8_t velocity;
 
 };
 #endif
